Finalised results page and added all sprites
